api/requirements.txt
Flask==1.1.1
PyMySQL==0.9.3
requests==2.23.0
SQLAlchemy==1.3.15
flask_cors==3.0.8
pycryptodome==3.7.2
pycryptodomex==3.6.4
